+++ src/body/CollisionBody.cpp
@@ -117,11 +117,11 @@ Collider* CollisionBody::addCollider(CollisionShape* collisionShape, const Trans
     mWorld.mCollisionDetection.addCollider(collider, aabb);
 
     RP3D_LOG(mLogger, Logger::Level::Information, Logger::Category::Body,
-             "Body " + std::to_string(mEntity.id) + ": Collider " + std::to_string(collider->getBroadPhaseId()) + " added to body");
+             "Body " + std::to_string(mEntity.id) + ": Collider " + std::to_string(collider->getBroadPhaseId()) + " added to body",  __FILE__, __LINE__);
 
     RP3D_LOG(mLogger, Logger::Level::Information, Logger::Category::Collider,
              "Collider " + std::to_string(collider->getBroadPhaseId()) + ":  collisionShape=" +
-             collider->getCollisionShape()->to_string());
+             collider->getCollisionShape()->to_string(),  __FILE__, __LINE__);
 
     // Return a pointer to the collision shape
     return collider;
@@ -171,7 +171,7 @@ Collider* CollisionBody::getCollider(uint colliderIndex) {
 void CollisionBody::removeCollider(Collider* collider) {
 
     RP3D_LOG(mLogger, Logger::Level::Information, Logger::Category::Body,
-             "Body " + std::to_string(mEntity.id) + ": Collider " + std::to_string(collider->getBroadPhaseId()) + " removed from body");
+             "Body " + std::to_string(mEntity.id) + ": Collider " + std::to_string(collider->getBroadPhaseId()) + " removed from body",  __FILE__, __LINE__);
 
     // Remove the collider from the broad-phase
     if (collider->getBroadPhaseId() != -1) {
@@ -283,7 +283,7 @@ void CollisionBody::setIsActive(bool isActive) {
 
     RP3D_LOG(mLogger, Logger::Level::Information, Logger::Category::Body,
              "Body " + std::to_string(mEntity.id) + ": Set isActive=" +
-             (isActive ? "true" : "false"));
+             (isActive ? "true" : "false"),  __FILE__, __LINE__);
 }
 
 // Ask the broad-phase to test again the collision shapes of the body for collision
@@ -399,7 +399,7 @@ void CollisionBody::setTransform(const Transform& transform) {
     updateBroadPhaseState(0);
 
     RP3D_LOG(mLogger, Logger::Level::Information, Logger::Category::Body,
-             "Body " + std::to_string(mEntity.id) + ": Set transform=" + transform.to_string());
+             "Body " + std::to_string(mEntity.id) + ": Set transform=" + transform.to_string(),  __FILE__, __LINE__);
 }
 
 // Return true if the body is active